requirement.txt
google-cloud-aiplatform==1.25.0
scikit-learn==1.2.2
pandas==2.0.2
#pyyaml==6.0
joblib==1.2.0
xgboost==1.7.5
lightgbm==3.3.5